backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
pytest==7.4.3
python-jose==3.3.0
passlib==1.7.4
